feat: Cleanup some German streaming platforms
ProSiebenSat1 (German TV production company) initially owned both maxdome and Joyn. The sold maxdome and moved the content to Joyn.

Date-A-Live IV will be the first ever show to be broadcast in Japanese with German subtitles as a simulcast on German TV (We usually only have dubs for everything, even foreign Reality-TV shows). The episodes will then be available on Joyn. Everything else is dubs only there.